+++ examples/aditi/aditi/settings.py
@@ -39,9 +39,6 @@ INSTALLED_APPS = [ # should change to first_app to get the app.
     "navycut.contrib.admin",
     "navycut.helpers.upload_server",
     "aniket",
-    "sarkar",
-    "kolkata",
-    "plassey"
 ]
 
 MIDDLEWARE = [