+++ examples/aditi/aniket/models.py
@@ -7,17 +7,20 @@ from navycut.orm import sql
 class Blog(sql.Model):
     # id = sql.fields.Integer(pk=True, unique=True)
     name = sql.fields.Char(required=True, unique=True)
-    picture = sql.fields.Image(required=True)
+    picture = sql.fields.Image()
     body = sql.fields.Text(required=False)
     author_id = sql.fields.ForeignKey("Author")
-    author = sql.fields.ManyToOne("Author", backref="blogs")
+    author = sql.fields.Relationship("Author")
     is_active = sql.fields.Boolean(default=True)
 
 class Author(sql.Model):
     # id = sql.fields.Integer(pk=True, unique=True)
     name = sql.fields.Char(required=True, unique=True)
-    picture = sql.fields.Image(required=False)
+    picture = sql.fields.Image()
     is_active = sql.fields.Boolean(default=True)
+
+    def __repr__(self):
+        return self.name
 
 class ExtraUser(sql.Model):
     # id = sql.fields.Integer(pk=True, unique=True)