+++ examples/aditi/kolkata/admin.py
@@ -1,4 +1,0 @@
-from navycut.contrib.admin import admin
-from .models import *
-
-admin.register_model(Blog1)