+++ examples/aditi/kolkata/urls.py
@@ -1,6 +1,0 @@
-from navycut.urls import path
-# from . import views
-
-url_patterns = [
-    # path("" , views.IndexView, "index"),
-]