+++ examples/aditi/sarkar/views.py
@@ -1,8 +1,0 @@
-from navycut.urls import MethodView
-# from navycut.http import JsonResponse
-
-# write your views here.
-
-# class IndexView(MethodView):
-#     def get(self):
-#         return JsonResponse(message="Salve Mundi!")