+++ examples/check2/nord/admin.py
@@ -1,4 +1,0 @@
-from navycut.admin import admin
-from .models import *
-
-admin.register_model(AniketSarkar, category=str(__name__).split(".")[0].capitalize())