+++ examples/dip/blog/admin.py
@@ -1,5 +1,0 @@
-from navycut.admin import admin
-from .models import *
-
-admin.register_model(Gargi)
-admin.register_model(Aniket)