+++ examples/dip/blog/urls.py
@@ -1,7 +1,0 @@
-from navycut.urls import path
-from .views import *
-
-url_patterns = [
-    path("" , IndexView, "index"),
-    path("gargi", GargiView, "gargi"),
-]