+++ examples/dip/dip/wsgi.py
@@ -1,6 +1,0 @@
-from . import settings
-
-settings.app.add_config(settings)
-settings.app.debugging(False)
-
-application = settings.app