+++ examples/dip/manage.py
@@ -1,4 +1,0 @@
-if __name__ == '__main__':
-    from navycut.command import Command
-    command:Command = Command()
-    command.run()